// src/app/shared/types/api-types/bire-alternative-pn-and-item-pn-dto.interface.ts
export interface BireAlternativePnAndItemPnDTO {
  pnPnCode?: string;
  pnCode?: string;
  familyCode?: string;
  structureType?: string;
  variantCode?: string;
  chapter?: string;
  sheet?: string;
  marks?: string;
  section?: string;
  subject?: string;
  versionNumber?: string;
  interchangeability?: string;
  statusState?: string;
  statusUser?: string;
  statusDate?: Date;
  quantityBefore?: string;
  quantityAfter?: string;
  actionTypeCode?: string;
  evolutionNumber?: string;
  evolutionRevisionNumber?: string;
  evolutionRemarks?: string;
  applicability?: string;
}