// src/app/shared/types/api-types/bire-pn-dto.interface.ts
export interface BirePnDTO {
  pnCode?: string;
  pnType?: string;
  manufacturerOtanCode?: string;
  otanPnReference?: string;
  fromReparationSolution?: boolean;
  quantityUnit?: string;
  articleDesignation?: string;
  additionalText?: string;
  pnCode1?: string;
  pnCode2?: string;
  pnCode3?: string;
  pnCode4?: string;
  pnCode5?: string;
  pnCode6?: string;
  manufacturerOtanCode1?: string;
  manufacturerOtanCode2?: string;
  manufacturerOtanCode3?: string;
  manufacturerOtanCode4?: string;
  manufacturerOtanCode5?: string;
  manufacturerOtanCode6?: string;
  isPma?: boolean;
  traceability?: string;
  expirationDuration?: string;
  expirationUnit?: string;
  expirationThreshold?: string;
  expirationThresholdUnit?: string;
  unitOfIssue?: string;
  quantityPerIssue?: string;
  statusState?: string;
  statusUser?: string;
  statusDate?: Date;
  incompletePartForSupply?: boolean;
  averagePrice?: number;
  sparePartClassCode?: string;
  unpackedHeight?: string;
  unpackedLength?: string;
  unpackedWidth?: string;
  packedWeight?: string;
  netWeight?: string;
  mtbf?: string;
  mtbur?: string;
  accounting?: string;
  hasExpirationDate?: boolean;
}